+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Add ``shift()`` method to the ``time_machine`` pytest fixture.
+
+  Thanks to Stefaan Lippens in `PR 312 <https://github.com/adamchainz/time-machine/pull/312>`__.
+
 * Mock ``time.monotonic()`` and ``time.monotonic_ns()``.
   They return the values of ``time.time()`` and ``time.time_ns()`` respectively, rather than real monotonic clocks.